requirements.txt
pandas==1.2.4
numpy==1.20.3
streamlit==0.82.0
Pillow==8.2.0
